+++ CHANGELOG.md
@@ -1,3 +1,124 @@
+#### 0.0.6 (2020-12-12)
+
+##### Build System / Dependencies
+
+* **deps:**  Bump ini from 1.3.5 to 1.3.7 ([c2538f30](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/c2538f3012ab6b9dda1009bb36dcd4eab6ab2a2e))
+* **deps-dev:**
+  *  Bump @babel/register from 7.12.1 to 7.12.10 ([39b77a61](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/39b77a61210291920cebbe53c820fb94a2696ce1))
+  *  Bump @types/node from 14.14.11 to 14.14.12 ([aca30d1c](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/aca30d1cd57a7921e8d4eb8dd3f4aac5287c83c9))
+  *  Bump @babel/core from 7.12.9 to 7.12.10 ([85b149d7](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/85b149d70287e2db2f90d377a6b3c38071741d35))
+  *  Bump @typescript-eslint/eslint-plugin ([3e374dd4](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/3e374dd4cc2a4c4fccfe2ecefd1b580ab6c530be))
+  *  Bump @types/node from 14.14.6 to 14.14.11 ([152e1ac1](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/152e1ac1fda196096af17751fadb76d77b87c9d9))
+  *  Bump @typescript-eslint/parser from 4.6.1 to 4.9.1 ([6b5853c2](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/6b5853c2a7d85b6759adf3a215226947983c9350))
+  *  Bump eslint from 7.12.1 to 7.15.0 ([e84199c8](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/e84199c84aeb9883a4dcf460b6332b310564d427))
+  *  Bump gulp-sourcemaps from 2.6.5 to 3.0.0 ([8883a7e3](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/8883a7e34872dd71a32827f6036a099b7ab302e7))
+  *  Bump @typescript-eslint/eslint-plugin ([908f6e47](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/908f6e471e7e84a7bd2bcb61433ab47cfc579ab0))
+  *  Bump @typescript-eslint/parser from 4.6.0 to 4.6.1 ([89377a4f](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/89377a4f31dbd0507713c387e1c541297b23f8d6))
+
+##### New Features
+
+*  Spell Sheet Done. (For Now.) ([76288b70](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/76288b70a5ff06310fea5a62dd455ce7ebd38b7e))
+*  Weapon Sheet Updates. ([ce7ff005](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/ce7ff005767156d83eb4e8c3b024493ce964b2d3))
+*  Template Partials. ([26386552](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/263865527262a56e06480e3c93da59714d5454c8))
+*  More Config Options. ([d29d4d15](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/d29d4d154fe0feff2feb6858dca675c16d26c178))
+*  Template Updates. ([0172bc7b](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/0172bc7bc31023a942ceaa58995ddaf3fdd7e274))
+*  Migrations and Template Pre-loading. ([0bc194d3](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/0bc194d37f7d3adea0af845fc5322861a87e71c7))
+*  Enchantment Sheet Done. (For Now.) ([a13a8882](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/a13a8882b8068670f6bec4a8141647a66b2b40ae))
+*  Armor Sheet Done. (For Now.) ([7fd3adec](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/7fd3adec98b6c6202c88732db4acb8fbd91471e1))
+*  Added Config Options. ([c61f08ad](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/c61f08adaa143f8cf1951eaebabf818b0ded461b))
+*  Weapon Sheet Done (For now.) ([83c85d69](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/83c85d691519d5ca576381cb96a4ed42c977d9a9))
+*  Weapon Header and Nav Html. ([23090b88](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/23090b8886e123b769003b6882232c8484006831))
+*  Added more Handlebar Helpers. ([a0ff1bcf](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/a0ff1bcf95b91976ec4d3c1e04299a9c60107605))
+*  Added Mod Calc Handlebar Helper ([6b7ea3b7](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/6b7ea3b7b2353756ccd47cfd59f00e8b87b8bd76))
+*  Added Images to Repo. ([ec739ddf](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/ec739ddf1f08aa0d3d169e8c9bb4defb3bc51a72))
+*  Prepped Remaining Item Sheets. ([3b9a8ed6](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/3b9a8ed62110aec2121ef96ad5f5591b9288ce2e))
+*  Added new Templates. ([9660dfae](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/9660dfae7639fc6a29dd34bf5db9bcc901fdfc16))
+
+##### Bug Fixes
+
+*  Typo. ([d1adb964](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/d1adb964efd65289293eeb405cc5f44a01ce9c04))
+*  Fixed Helpers. ([f3bddb52](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/f3bddb528c66e6bce3683183968836c45b7869d7))
+*  Fixed Other Sheets. ([e776f3c5](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/e776f3c53ac8d076084f5fb0f7dfa4767e038022))
+*  Localization and Navigation fixes. ([b00589d3](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/b00589d3df69cc2232dd108ec54e97bcd62fb5c7))
+*  Update System.json ([754e3051](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/754e305164a4cbd10bb35aed5fa298f90e8dcaea))
+
+##### Other Changes
+
+*  Console Log Removal. ([989d9423](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/989d94232b10f630ba13fdda7710a66d432012a9))
+*  More Localization Strings. ([bac357ad](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/bac357ada120467f3a9690d43acff0dabd16c272))
+*  Even More Strings. ([22ff0b3f](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/22ff0b3ff3a65125648ade75ae6cdab084306c5b))
+*  Added More Strings. ([bced986c](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/bced986c300e9ca220526771d1fb8069b9fc5e9f))
+*  Added Template Data. ([ab160570](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/ab16057087820120e285c7ff2f30a0ef736cb94c))
+*  Placeholder Strings ([a0c62a5d](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/a0c62a5de1f624490d66c06a1c25d7a186dbfad9))
+*  Added Forgotten String ([5c399f13](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/5c399f13aaaabdd386bf5879e1b04b7d55a772d4))
+*  Added New Strings, copied English to Spanish and Italian. ([8258d4f1](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/8258d4f1492b97aac7792810980a928fe583b7fa))
+*  Edited Templates. ([25546bc0](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/25546bc0843a7b7fe9e23adc14257138c8d58474))
+*  Added Build Status Tags. ([a57cb01b](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/a57cb01b00301eefd0b2055b61b844d22875c50a))
+*  Added new Localization strings. ([269b57a3](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/269b57a3649bf94c491df60497390ab3d6c47cb3))
+*  Update node-sass ([4972b07a](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/4972b07a5d83df9d01adb6c9bb615d01d18ce03d))
+
+##### Code Style Changes
+
+*  Style Fixes. ([dbf5d2d3](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/dbf5d2d3b0aafda8fa6ed79ffc5d6fdf085ed27c))
+*  Header and Nav Styles for Weapons, Reorg of Hero Styles. ([78cd3776](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/78cd3776984b0c2afc72bba5c5df932f09f03a0b))
+*  Restyle existing Sheet Styles with DnD5e CSS. ([3dcd1cb9](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/3dcd1cb9640a5d1415e22bcbd2030cc1f011e64f))
+
+#### 0.0.5 (2020-12-11)
+
+##### Build System / Dependencies
+
+* **deps:**  Bump ini from 1.3.5 to 1.3.7 ([c2538f30](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/c2538f3012ab6b9dda1009bb36dcd4eab6ab2a2e))
+* **deps-dev:**
+  *  Bump @babel/register from 7.12.1 to 7.12.10 ([39b77a61](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/39b77a61210291920cebbe53c820fb94a2696ce1))
+  *  Bump @types/node from 14.14.11 to 14.14.12 ([aca30d1c](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/aca30d1cd57a7921e8d4eb8dd3f4aac5287c83c9))
+  *  Bump @babel/core from 7.12.9 to 7.12.10 ([85b149d7](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/85b149d70287e2db2f90d377a6b3c38071741d35))
+  *  Bump @typescript-eslint/eslint-plugin ([3e374dd4](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/3e374dd4cc2a4c4fccfe2ecefd1b580ab6c530be))
+  *  Bump @types/node from 14.14.6 to 14.14.11 ([152e1ac1](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/152e1ac1fda196096af17751fadb76d77b87c9d9))
+  *  Bump @typescript-eslint/parser from 4.6.1 to 4.9.1 ([6b5853c2](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/6b5853c2a7d85b6759adf3a215226947983c9350))
+  *  Bump eslint from 7.12.1 to 7.15.0 ([e84199c8](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/e84199c84aeb9883a4dcf460b6332b310564d427))
+  *  Bump gulp-sourcemaps from 2.6.5 to 3.0.0 ([8883a7e3](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/8883a7e34872dd71a32827f6036a099b7ab302e7))
+  *  Bump @typescript-eslint/eslint-plugin ([908f6e47](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/908f6e471e7e84a7bd2bcb61433ab47cfc579ab0))
+  *  Bump @typescript-eslint/parser from 4.6.0 to 4.6.1 ([89377a4f](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/89377a4f31dbd0507713c387e1c541297b23f8d6))
+
+##### New Features
+
+*  Enchantment Sheet Done. (For Now.) ([a13a8882](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/a13a8882b8068670f6bec4a8141647a66b2b40ae))
+*  Armor Sheet Done. (For Now.) ([7fd3adec](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/7fd3adec98b6c6202c88732db4acb8fbd91471e1))
+*  Added Config Options. ([c61f08ad](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/c61f08adaa143f8cf1951eaebabf818b0ded461b))
+*  Weapon Sheet Done (For now.) ([83c85d69](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/83c85d691519d5ca576381cb96a4ed42c977d9a9))
+*  Weapon Header and Nav Html. ([23090b88](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/23090b8886e123b769003b6882232c8484006831))
+*  Added more Handlebar Helpers. ([a0ff1bcf](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/a0ff1bcf95b91976ec4d3c1e04299a9c60107605))
+*  Added Mod Calc Handlebar Helper ([6b7ea3b7](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/6b7ea3b7b2353756ccd47cfd59f00e8b87b8bd76))
+*  Added Images to Repo. ([ec739ddf](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/ec739ddf1f08aa0d3d169e8c9bb4defb3bc51a72))
+*  Prepped Remaining Item Sheets. ([3b9a8ed6](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/3b9a8ed62110aec2121ef96ad5f5591b9288ce2e))
+*  Added new Templates. ([9660dfae](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/9660dfae7639fc6a29dd34bf5db9bcc901fdfc16))
+
+##### Bug Fixes
+
+*  Typo. ([d1adb964](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/d1adb964efd65289293eeb405cc5f44a01ce9c04))
+*  Fixed Helpers. ([f3bddb52](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/f3bddb528c66e6bce3683183968836c45b7869d7))
+*  Fixed Other Sheets. ([e776f3c5](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/e776f3c53ac8d076084f5fb0f7dfa4767e038022))
+*  Localization and Navigation fixes. ([b00589d3](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/b00589d3df69cc2232dd108ec54e97bcd62fb5c7))
+*  Update System.json ([754e3051](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/754e305164a4cbd10bb35aed5fa298f90e8dcaea))
+
+##### Other Changes
+
+*  Even More Strings. ([22ff0b3f](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/22ff0b3ff3a65125648ade75ae6cdab084306c5b))
+*  Added More Strings. ([bced986c](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/bced986c300e9ca220526771d1fb8069b9fc5e9f))
+*  Added Template Data. ([ab160570](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/ab16057087820120e285c7ff2f30a0ef736cb94c))
+*  Placeholder Strings ([a0c62a5d](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/a0c62a5de1f624490d66c06a1c25d7a186dbfad9))
+*  Added Forgotten String ([5c399f13](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/5c399f13aaaabdd386bf5879e1b04b7d55a772d4))
+*  Added New Strings, copied English to Spanish and Italian. ([8258d4f1](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/8258d4f1492b97aac7792810980a928fe583b7fa))
+*  Edited Templates. ([25546bc0](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/25546bc0843a7b7fe9e23adc14257138c8d58474))
+*  Added Build Status Tags. ([a57cb01b](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/a57cb01b00301eefd0b2055b61b844d22875c50a))
+*  Added new Localization strings. ([269b57a3](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/269b57a3649bf94c491df60497390ab3d6c47cb3))
+*  Update node-sass ([4972b07a](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/4972b07a5d83df9d01adb6c9bb615d01d18ce03d))
+
+##### Code Style Changes
+
+*  Header and Nav Styles for Weapons, Reorg of Hero Styles. ([78cd3776](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/78cd3776984b0c2afc72bba5c5df932f09f03a0b))
+*  Restyle existing Sheet Styles with DnD5e CSS. ([3dcd1cb9](https://github.com/HeikaHaku/fvtt-carrot-royale/commit/3dcd1cb9640a5d1415e22bcbd2030cc1f011e64f))
+
 #### 0.0.4 (2020-12-11)
 
 ##### Build System / Dependencies